enhance chat module: refactor package structure for ChatRepository, DoctorChatScreen, and Message
